cut up the sign component with extra error notifictions
